# backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ALN Orchestrator API
  version: 1.0.0
  description: |
    About Last Night (ALN) Orchestrator backend API for memory token scanning system.

    **Architecture**: Contract-First approach with minimal complexity for live-event tool.

    **Components**:
    - Player Scanner: HTTP-only, fire-and-forget (ESP32 compatible)
    - GM Scanner: WebSocket-driven real-time game logic
    - Admin Panel: Monitoring and intervention via WebSocket

    **Key Constraints**:
    - ONE session at a time (2-hour live event)
    - Standalone mode support (scanners work without orchestrator)
    - Minimal architectural complexity (not enterprise SaaS)

    **Documentation References**:
    - Alignment Decisions: docs/api-alignment/04-alignment-decisions.md
    - Functional Requirements: docs/api-alignment/08-functional-requirements.md
    - Test Architecture: docs/api-alignment/06-test-architecture.md
servers:
  - url: http://localhost:3000
    description: Local development
  - url: http://{orchestrator-ip}:3000
    description: Network deployment
    variables:
      orchestrator-ip:
        default: "10.0.0.100"
        description: Orchestrator IP address on local network
...
paths:
...
  /api/session:
    get:
      tags:
        - Session Management
      summary: Get current active session
      description: |
        Get current active session information (one-time fetch).

        **Functional Requirement**: Section 1.2 (Session Management) - ONE session at a time
        **Decisions Applied**:
        - #3 (RESTful HTTP - direct response, HTTP codes)
        - #4 (id field within resource, not sessionId)
        - #7 (full session resource structure)

        **CRITICAL - ONE Session at a Time**:
        This system supports ONLY ONE active session (2-hour live event model).
        Multi-session endpoints (GET /api/session/:id, GET /api/admin/sessions) ELIMINATED.

        **Usage**:
        - GM Scanner: One-time fetch on connection (before WebSocket established)
        - Debug/Recovery: Check current session state via HTTP
        - Primary State: Comes via WebSocket sync:full event (not HTTP polling)

        **Session Commands**:
        Session lifecycle commands (create/pause/resume/end) handled via WebSocket gm:command event, NOT HTTP POST endpoints.

        **Breaking Changes**:
        - Field: sessionId → id (Decision #4 - use 'id' within resource)
        - Added: teams array (target state, not currently implemented)
        - Structure: Full resource matches session:update WebSocket event (Decision #7)
      operationId: getCurrentSession
      responses:
        '200':
          description: Current session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
              example:
                id: "2a2f9d45-5d2d-441d-b32c-52c939f3c103"
                name: "About Last Night - Oct 15 2025"
                startTime: "2025-10-15T19:00:00.000Z"
                endTime: null
                status: "active"
                teams: ["Team Alpha", "Detectives", "Blue Squad"]
                metadata:
                  gmStations: 2
                  playerDevices: 3
                  totalScans: 47
                  uniqueTokensScanned: ["534e2b02", "534e2b03", "tac001", "fli001", "hos001"]
        '404':
          description: No active session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "SESSION_NOT_FOUND"
                message: "No active session"
  /api/state:
    get:
      tags:
        - Static Resources
      summary: Get current game state (debug/recovery)
      description: |
        Get complete current game state for debugging or recovery (one-time fetch).

        **Functional Requirement**: Section 1.7 (State Synchronization) - debugging utility
        **Decisions Applied**:
        - #3 (RESTful HTTP)
        - #5 (video:status structure - used in videoStatus field)

        **CRITICAL - NOT for Polling**:
        This endpoint is for DEBUGGING and RECOVERY only.
        Real-time state synchronization provided by WebSocket sync:full event.

        Investigation finding: No polling pattern detected (correct usage).

        **ETag Caching**:
        Supports HTTP ETag for efficient caching (304 Not Modified response).
        Client sends If-None-Match header with previous ETag.

        **State Structure**:
        Complete snapshot including:
        - Current session (if exists)
        - All team scores
        - Recent transactions (last 100)
        - Video playback status (matches video:status WebSocket event per Decision #5)
        - Connected devices
        - System health (orchestrator + VLC)

        **Usage**:
        - Debug: Manual state inspection during development
        - Recovery: Fetch state after disconnect/reconnect
        - Admin Panel: Initial state load before WebSocket connection
      operationId: getCurrentState
      parameters:
        - in: header
          name: If-None-Match
          schema:
            type: string
          description: ETag from previous response for caching
          example: '"d41d8cd98f00b204e9800998ecf8427e"'
      responses:
        '200':
          description: Current game state
          headers:
            ETag:
              schema:
                type: string
              description: Entity tag for caching
              example: '"d41d8cd98f00b204e9800998ecf8427e"'
            Cache-Control:
              schema:
                type: string
              description: Cache control header
              example: "no-cache, must-revalidate"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameState'
              example:
                session:
                  id: "2a2f9d45-5d2d-441d-b32c-52c939f3c103"
                  name: "About Last Night - Oct 15 2025"
                  startTime: "2025-10-15T19:00:00.000Z"
                  endTime: null
                  status: "active"
                  teams: ["Team Alpha", "Detectives", "Blue Squad"]
                  metadata:
                    gmStations: 2
                    playerDevices: 3
                    totalScans: 47
                scores:
                  - teamId: "Team Alpha"
                    currentScore: 11500
                    baseScore: 11000
                    bonusPoints: 500
                    tokensScanned: 8
                    completedGroups: ["jaw_group"]
                    lastUpdate: "2025-10-15T20:15:30.000Z"
                  - teamId: "Detectives"
                    currentScore: 9200
                    baseScore: 9200
                    bonusPoints: 0
                    tokensScanned: 6
                    completedGroups: []
                    lastUpdate: "2025-10-15T20:14:15.000Z"
                recentTransactions:
                  - id: "7b8b1d85-b234-4be9-bde5-4c8522a1f15e"
                    tokenId: "534e2b03"
                    teamId: "Team Alpha"
                    deviceId: "GM_Station_1"
                    mode: "blackmarket"
                    points: 3000
                    timestamp: "2025-10-15T20:15:30.000Z"
                    memoryType: "Technical"
                    valueRating: 3
                    group: "jaw_group"
                  - id: "419d5341-25e9-4ef4-a0a5-9c3a3ec19b2a"
                    tokenId: "tac001"
                    teamId: "Detectives"
                    deviceId: "GM_Station_2"
                    mode: "detective"
                    points: 100
                    timestamp: "2025-10-15T20:14:15.000Z"
                    memoryType: "Personal"
                    valueRating: 1
                    group: ""
                videoStatus:
                  status: "playing"
                  queueLength: 2
                  tokenId: "534e2b03"
                  duration: 30
                  progress: 45
                  expectedEndTime: "2025-10-15T20:16:00.000Z"
                  error: null
                devices:
                  - deviceId: "GM_Station_1"
                    type: "gm"
                    name: "GM Station v1.0.0"
                    connectionTime: "2025-10-15T19:00:30.000Z"
                    ipAddress: "10.0.0.81"
                  - deviceId: "PLAYER_SCANNER_01"
                    type: "player"
                    name: "Player Scanner 1"
                    connectionTime: "2025-10-15T19:02:00.000Z"
                    ipAddress: "10.0.0.82"
                serviceHealth:
                  vlc:
                    status: "healthy"
                    message: "Connected to VLC"
                    lastChecked: "2025-10-15T19:00:30.000Z"
                  spotify:
                    status: "down"
                    message: "Not yet checked"
                    lastChecked: null
        '304':
          description: Not Modified (ETag match)
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          enum:
            - AUTH_REQUIRED
            - PERMISSION_DENIED
            - VALIDATION_ERROR
            - SESSION_NOT_FOUND
            - TOKEN_NOT_FOUND
            - DUPLICATE_TRANSACTION
            - INVALID_REQUEST
        message:
          type: string
          description: User-friendly error message
        details:
          type: array
          description: Optional error details
          items:
            type: object
...
    Session:
      type: object
      required:
        - id
        - name
        - startTime
        - status
        - teams
        - metadata
      properties:
        id:
          type: string
          format: uuid
          description: Session ID (UUID v4, per Decision 4 - use 'id' within resource)
          example: "2a2f9d45-5d2d-441d-b32c-52c939f3c103"
        name:
          type: string
          description: Event name
          minLength: 1
          maxLength: 100
          example: "About Last Night - Oct 15 2025"
        startTime:
          type: string
          format: date-time
          description: Session start time (ISO8601)
          example: "2025-10-15T19:00:00.000Z"
        endTime:
          type: string
          format: date-time
          nullable: true
          description: Session end time (null if active)
          example: null
        status:
          type: string
          enum: [setup, active, paused, ended]
          description: Session status
          example: "active"
        teams:
          type: array
          items:
            type: string
            description: Team ID
          description: Array of team IDs participating in session
          example: ["Team Alpha", "Detectives", "Blue Squad"]
        metadata:
          type: object
          description: Session metadata (statistics, device counts)
          properties:
            gmStations:
              type: integer
              description: Number of GM stations connected
              example: 2
            playerDevices:
              type: integer
              description: Number of player scanners connected
              example: 3
            totalScans:
              type: integer
              description: Total scans across all teams
              example: 47
            uniqueTokensScanned:
              type: array
              items:
                type: string
              description: List of unique token IDs scanned
              example: ["534e2b02", "534e2b03", "tac001", "fli001", "hos001"]
    GameState:
      type: object
      required:
        - session
        - scores
        - recentTransactions
        - videoStatus
        - devices
        - serviceHealth
      properties:
        session:
          oneOf:
            - $ref: '#/components/schemas/Session'
            - type: 'null'
          description: Current active session (null if no session)
        scores:
          type: array
          description: Team scores array
          items:
            type: object
            required:
              - teamId
              - currentScore
              - baseScore
              - bonusPoints
              - tokensScanned
              - completedGroups
              - lastUpdate
            properties:
              teamId:
                type: string
                description: Team ID
                example: "Team Alpha"
              currentScore:
                type: integer
                description: Total score (base + bonus)
                example: 11500
              baseScore:
                type: integer
                description: Base score from token values
                example: 11000
              bonusPoints:
                type: integer
                description: Bonus points from group completions
                example: 500
              tokensScanned:
                type: integer
                description: Number of tokens scanned by team
                example: 8
              completedGroups:
                type: array
                items:
                  type: string
                description: Array of completed group IDs
                example: ["jaw_group"]
              lastUpdate:
                type: string
                format: date-time
                description: Last score update timestamp
                example: "2025-10-15T20:15:30.000Z"
        recentTransactions:
          type: array
          description: Recent transactions (last 100)
          items:
            type: object
            required:
              - id
              - tokenId
              - teamId
              - deviceId
              - mode
              - points
              - timestamp
            properties:
              id:
                type: string
                format: uuid
                description: Transaction ID
                example: "7b8b1d85-b234-4be9-bde5-4c8522a1f15e"
              tokenId:
                type: string
                description: Token ID (per Decision 4)
                example: "534e2b03"
              teamId:
                type: string
                description: Team ID
                example: "Team Alpha"
              deviceId:
                type: string
                description: GM Scanner device ID (per Decision 4)
                example: "GM_Station_1"
              mode:
                type: string
                enum: [detective, blackmarket]
                description: Scan mode
                example: "blackmarket"
              points:
                type: integer
                description: Points awarded (0 if duplicate/error)
                example: 3000
              timestamp:
                type: string
                format: date-time
                description: Transaction timestamp
                example: "2025-10-15T20:15:30.000Z"
              memoryType:
                type: string
                enum: [Technical, Business, Personal]
                description: Token memory type (from tokens.json)
                example: "Technical"
              valueRating:
                type: integer
                minimum: 1
                maximum: 5
                description: Token value rating (from tokens.json)
                example: 3
              group:
                type: string
                description: Token group (from tokens.json, empty string if no group)
                example: "jaw_group"
        videoStatus:
          type: object
          required:
            - status
            - queueLength
          description: Current video playback status (per Decision 5)
          properties:
            status:
              type: string
              enum: [idle, loading, playing, paused, completed, error]
              description: Video playback status
              example: "playing"
            queueLength:
              type: integer
              description: Number of videos in queue (per Decision 5 - ADDED FIELD)
              minimum: 0
              example: 2
            tokenId:
              type: string
              nullable: true
              description: Currently playing token ID (null if idle)
              example: "534e2b03"
            duration:
              type: number
              nullable: true
              description: Video duration in seconds (null if no video)
              example: 30
            progress:
              type: number
              nullable: true
              description: Playback progress percentage (0-100)
              minimum: 0
              maximum: 100
              example: 45
            expectedEndTime:
              type: string
              format: date-time
              nullable: true
              description: Expected completion time (null if no video)
              example: "2025-10-15T20:16:00.000Z"
            error:
              type: string
              nullable: true
              description: Error message if status=error
              example: null
        devices:
          type: array
          description: Connected devices
          items:
            type: object
            required:
              - deviceId
              - type
              - name
              - connectionTime
            properties:
              deviceId:
                type: string
                description: Device identifier (per Decision 4)
                example: "GM_Station_1"
              type:
                type: string
                enum: [gm, player]
                description: Device type
                example: "gm"
              name:
                type: string
                description: Friendly device name
                example: "GM Station v1.0.0"
              connectionTime:
                type: string
                format: date-time
                description: Connection timestamp
                example: "2025-10-15T19:00:30.000Z"
              ipAddress:
                type: string
                description: Device IP address
                example: "10.0.0.81"
        serviceHealth:
          type: object
          description: Health snapshot for all registered services from serviceHealthRegistry
          additionalProperties:
            type: object
            required:
              - status
              - message
            properties:
              status:
                type: string
                enum: [healthy, down]
              message:
                type: string
              lastChecked:
                type: string
                format: date-time
                nullable: true
          example:
            vlc:
              status: "healthy"
              message: "Connected to VLC"
              lastChecked: "2025-10-15T19:00:30.000Z"
            spotify:
              status: "down"
              message: "Not yet checked"
              lastChecked: null
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from POST /api/admin/auth